CLIM-1447: Atom4 make obvious apps/src/components/s2d/... -- MOVING-THINGS-AROUND UNDO
