+++ apps/src/components/download/step-variable-options.tsx
@@ -6,7 +6,7 @@ import {
 } from '@/components/download/step-container';
 import { AnalyzedDownloadFields } from "@/components/download/ui/analyzed-download-fields";
 import { VersionDownloadFields } from "@/components/download/ui/version-download-fields";
-import { S2DForecastTypeFieldDropdown } from '@/components/s2d/forecast';
+import { S2DForecastTypeFieldDropdown } from '@/components/fields/forecast';
 import { DefinitionList, type DefinitionItem } from '@/components/ui/definition-list';
 import { useClimateVariable } from "@/hooks/use-climate-variable";
 import { useS2D } from '@/hooks/use-s2d';
